Module test_utils::authority

Structs

SuiNode

SuiNodeHandle

Wrap SuiNode to allow correct access to SuiNode in simulator tests.

Constants

NETWORK_BUFFER_SIZE

The default network buffer size of a test authority.

Functions

get_client

Get a network client to communicate with the consensus.

get_object

spawn_fullnode

This function can be called after spawn_test_authorities to start a fullnode.

spawn_test_authorities

Spawn all authorities in the test committee into a separate tokio task.

start_node

test_and_configure_authority_configs

test_authority_configs

Make an authority config for each of the TEST_COMMITTEE_SIZE authorities in the test committee.

test_authority_configs_with_objects