Module test_utils::messages

Constants

MAX_GAS

The maximum gas per transaction.

Functions

create_publish_move_package_transaction

Make a transaction to publish a test move contracts package.

get_account_and_gas_coins

A helper function to get all accounts and their owned GasCoin with a WalletContext

get_account_and_gas_objects

A helper function to get all accounts and their owned gas objects with a WalletContext.

get_gas_object_with_wallet_context

get one available gas ObjectRef

get_gas_objects_with_wallet_context

make_counter_create_transaction

make_counter_increment_transaction

make_counter_increment_transaction_with_wallet_context

make_publish_basics_transaction

make_random_certified_transaction

make_transactions_with_pre_genesis_objects

Make a few different single-writer test transactions owned by specific addresses.

make_transactions_with_wallet_context

A helper function to make Transactions with controlled accounts in WalletContext. Particularly, the wallet needs to own gas objects for transactions. However, if this function is called multiple times without any “sync” actions on gas object management, txns may fail and objects may be locked.

make_transfer_object_transaction_with_wallet_context

make_tx_certs_and_signed_effects

Make a test certificates for each input transaction.

make_tx_certs_and_signed_effects_with_committee

Make a test certificates for each input transaction.

move_transaction

Make a transaction calling a specific move module & function.

move_transaction_with_type_tags

Make a transaction calling a specific move module & function, with specific type tags

random_object_digest

test_shared_object_transactions

Make a few different test transaction containing the same shared object.