Struct test_utils::network::TestCluster

pub struct TestCluster {
    pub swarm: Swarm,
    pub wallet: WalletContext,
    pub fullnode_handle: FullNodeHandle,
}

Fields
swarm: Swarmwallet: WalletContextfullnode_handle: FullNodeHandle
Implementations

impl TestCluster

pub fn rpc_client(&self) -> &HttpClient

pub fn sui_client(&self) -> &SuiClient

pub fn rpc_url(&self) -> &str

pub fn wallet_mut(&mut self) -> &mut WalletContext

pub fn get_addresses(&self) -> Vec<SuiAddress>

pub fn get_address_0(&self) -> SuiAddress

pub fn get_address_1(&self) -> SuiAddress

pub fn get_address_2(&self) -> SuiAddress

pub fn fullnode_config_builder(&self) -> FullnodeConfigBuilder

pub async fn start_fullnode(&mut self) -> FullNodeHandle

Convenience method to start a new fullnode in the test cluster.

pub async fn start_fullnode_from_config( &mut self, config: NodeConfig ) -> FullNodeHandle

pub fn all_node_handles(&self) -> impl Iterator<Item = SuiNodeHandle>

pub fn get_validator_addresses(&self) -> Vec<AuthorityName>

pub fn stop_validator(&self, name: AuthorityName)

pub async fn stop_all_validators(&self)

pub async fn start_all_validator(&self)

pub async fn start_validator(&self, name: AuthorityName)

pub fn random_node_restarter(self: &Arc<Self>) -> RandomNodeRestarter

pub async fn get_reference_gas_price(&self) -> u64

pub async fn get_object_from_fullnode_store( &self, object_id: &ObjectID ) -> Option<Object>

pub async fn wait_for_epoch( &self, target_epoch: Option<EpochId> ) -> SuiSystemState

To detect whether the network has reached such state, we use the fullnode as the source of truth, since a fullnode only does epoch transition when the network has done so. If target_epoch is specified, wait until the cluster reaches that epoch. If target_epoch is None, wait until the cluster reaches the next epoch. Note that this function does not guarantee that every node is at the target epoch.

pub async fn wait_for_epoch_with_timeout( &self, target_epoch: Option<EpochId>, timeout_dur: Duration ) -> SuiSystemState

pub async fn trigger_reconfiguration(&self)

Ask 2f+1 validators to close epoch actively, and wait for the entire network to reach the next epoch. This requires waiting for both the fullnode and all validators to reach the next epoch.

pub async fn wait_for_epoch_all_validators(&self, target_epoch: EpochId)

pub async fn update_validator_supported_versions( &mut self, new_supported_versions: SupportedProtocolVersions )

Upgrade the network protocol version, by restarting every validator with a new supported versions. Note that we don’t restart the fullnode here, and it is assumed that the fulnode supports the entire version range.

pub async fn wait_for_all_nodes_upgrade_to(&self, protocol_version: u64)

Wait for all nodes in the network to upgrade to protocol_version.

pub async fn execute_transaction( &self, transaction: VerifiedTransaction ) -> SuiRpcResult<SuiTransactionBlockResponse>