Module test_utils::socket_utils

Enums

SockAddr

A container for different types of socket addresses.

SocketInitMethod

Functions

autobind_helper

A helper function to autobind the socket to some unused address. Returns the address.

dgram_connect_helper

A helper function to connect the client socket to the server address.

socket_init_helper

A helper function to initialize two sockets such that the first socket is connected to the second socket. The second socket may or may not be connected to the first socket.

stream_connect_helper

A helper function to connect the client socket to the server address. Returns the accepted socket.