Struct test_utils::Term

pub struct Term { /* fields omitted */ }

A wrapper around a termination signal [oneshot::Receiver].

This type manually implements Future in order to be able to be passed to a closure as part of background.