Crate test_utils

Utilities helpful for writing Rust integration tests.

Modules

socket_utils

Macros

check_system_call

Calls check_system_call(), but automatically passes the current line number.

ensure_ord

Convenience wrapper around anyhow::ensure that generates useful error messages.

set

Similar to the vec! macro, set! will create a HashSet with the given elements.

Structs

Interruptor

ShadowTest

Enums

TestEnvironment

Traits

AsMutPtr

An object that can be converted to a mutable pointer (possibly null).

AsPtr

An object that can be converted to a pointer (possibly null).

Functions

assert_true_else_errno

Assert the boolean condition is true, else print the last system error

check_system_call

Run the given function, check that the errno was expected, and return the function’s return value.

get_errno

Get the current errno.

get_errno_message

Get the message for the given errno.

i8_to_u8_slice

Convert a &[i8] to &[u8]. Useful when making C syscalls.

is_readable

Returns true if the POLLIN flag is set.

is_writable

Returns true if the POLLOUT flag is set.

nop_sig_handler

poll_status

Returns the poll event flags (the result of poll() with PollFlags::all() flags set). The flags will be empty if the timeout occurred.

result_assert

Return the error message if the condition is false.

result_assert_eq

Return a formatted error message if a and b are unequal.

result_assert_ne

Return a formatted error message if a and b are equal.

run_and_close_fds

Run the function and then close any given file descriptors, even if there was an error.

run_tests

Runs provided tests until failure and outputs results to stdout.

running_in_shadow

u8_to_i8_slice

Convert a &[u8] to &[i8]. Useful when making C syscalls.