Crate test_utils

Modules

marks
This module implements manually tracked test coverage, which useful for quickly finding a test responsible for testing a particular bit of code.

Macros

assert_eq_text

covers

marks

tested_by

Structs

FixtureEntry

__Changeset
The information about a full changeset

Constants

CURSOR_MARKER

Functions

add_cursor

collect_tests

dir_tests

extract_offset

extract_range

extract_ranges
Extracts ranges, marked with <tag> </tag> paris from the text

find_mismatch

lines_match
Compare a line with an expected pattern.

parse_fixture
Parses text which looks like this:

project_dir

read_text
Read file and normalize newlines.

try_extract_offset

try_extract_range