Crate test_utils

Functions

construct_block
Returns a full block used for testing

construct_chain_exchange_response
Returns a RPCResponse used for testing

construct_dummy_header

construct_epost_proof
Returns a deterministic EPostProof to be used for testing

construct_full_tipset
Returns a full tipset used for testing

construct_headers
Returns a vec of block headers to be used for testing purposes

construct_keys
Returns a vec of 4 distinct CIDs

construct_messages
Returns a tuple of unsigned and signed messages used for testing

construct_ticket
Returns a Ticket to be used for testing

construct_tipset
Returns a tipset used for testing

construct_tipset_bundle
Returns a TipsetBundle used for testing

template_key
Defines a TipsetKey used in testing