Module test_utils::transaction

Functions

compile_basics_package

create_devnet_nft

delete_devnet_nft

extract_obj

get_framework_object

Get the framework object

get_unique_effects

increment_counter

make_publish_package

parse_package_ref

Extract the package reference from a transaction effect. This is useful to deduce the authority-created package reference after attempting to publish a new Move package.

publish_basics_package

Helper function to publish basic package.

publish_basics_package_and_make_counter

A helper function to publish the basics package and make counter objects

publish_counter_package

Helper function to publish the move package of a simple shared counter.

publish_package

publish_package_for_effects

publish_package_with_wallet

Returns the published package’s ObjectRef.

split_coin_with_wallet_context

submit_move_transaction

A helper function to submit a move transaction

submit_shared_object_transaction

Keep submitting the certificates of a shared-object transaction until it is sequenced by at least one consensus node. We use the loop since some consensus protocols (like Tusk) may drop transactions. The certificate is submitted to every Sui authority.

submit_shared_object_transaction_with_committee

Keep submitting the certificates of a shared-object transaction until it is sequenced by at least one consensus node. We use the loop since some consensus protocols (like Tusk) may drop transactions. The certificate is submitted to every Sui authority.

submit_single_owner_transaction

Submit a certificate containing only owned-objects to all authorities.

transfer_coin

transfer_sui

wait_for_all_txes

wait_for_tx