Module test_utils::transaction

Functions

parse_package_ref

Extract the package reference from a transaction effect. This is useful to deduce the authority-created package reference after attempting to publish a new Move package.

publish_counter_package

Helper function to publish the move package of a simple shared counter.

publish_package

submit_shared_object_transaction

Keep submitting the certificates of a shared-object transaction until it is sequenced by at least one consensus node. We use the loop since some consensus protocols (like Tusk) may drop transactions. The certificate is submitted to every Sui authority.

submit_single_owner_transaction

Submit a certificate containing only owned-objects to all authorities.

wait_for_all_txes

wait_for_tx